server/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pyserial==3.5
cryptography==41.0.1
python-dotenv==1.0.0
pydantic==2.5.1